AppSentinel/AppScore/backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
Flask-CORS==3.0.10
psycopg2-binary==2.9.1
python-dotenv==0.19.0
alembic==1.7.1
SQLAlchemy==1.4.23
PyJWT==2.1.0
pytest==6.2.5
pytest-cov==2.12.1
Werkzeug==2.0.3
fpdf2==2.7.5
pandas==1.5.3
boto3==1.34.7
botocore==1.34.7
